댓글 삭제시 인가 처리, redirect 처리

+++ src/main/java/com/example/winku/controller/comment/CommentController.java
@@ -2,13 +2,19 @@ package com.example.winku.controller.comment;
 
 import com.example.winku.Service.comment.CommentServiceImpl;
 import com.example.winku.Service.recomment.RecommentServiceImpl;
+import com.example.winku.domain.user.User;
 import com.example.winku.dto.comment.CreateCommentDto;
 import com.example.winku.dto.comment.DeleteCommentDto;
 import com.example.winku.dto.recomment.CreateRecommentDto;
 import com.example.winku.dto.recomment.DeleteRecommentDto;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.SessionAttribute;
+import org.springframework.web.server.ResponseStatusException;
+
 @Controller
 public class CommentController {
     final CommentServiceImpl commentService;
@@ -31,15 +37,21 @@ public class CommentController {
     }
 
     @PostMapping("/feed/deleteComment")
-    public String CommentDelete(@ModelAttribute DeleteCommentDto commentDto) {
-        if(!commentService.findAllbyFeedId())
+    public String CommentDelete(@ModelAttribute DeleteCommentDto commentDto, @SessionAttribute(name = "user") User user, @RequestParam String redirect) {
+        if (!commentService.findbyCommentId(commentDto.getId()).getLoginId().equals(user.getLoginId())) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
         commentService.deleteComment(commentDto);
-        return "redirect:/feed/index";
+        return "redirect:" + redirect;
     }
+
     @PostMapping("/feed/deleteRecomment")
-    public String RecommentDelete(@ModelAttribute DeleteRecommentDto recommentDto) {
+    public String RecommentDelete(@ModelAttribute DeleteRecommentDto recommentDto, @SessionAttribute(name = "user") User user, @RequestParam String redirect) {
+        if (!recommentService.findbyRecommentId(recommentDto.getId()).getLoginId().equals(user.getLoginId())) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
         recommentService.deleteRecomment(recommentDto);
 
-        return "redirect:/feed/index";
+        return "redirect:" + redirect;
     }
 }
